Fix summary info
This fixes the computation and display of the summary info, so we have basically a reasonable `--dry-run` version of the script! There are several comments about features that really should be controlled by flags, and I think there are places where functions would be useful, but we've made significant headway.